refactor(gateway): extract turn-process supervision helpers from run.py (slice 10 of #54962)
